pyproject.toml
[project]
name = "flare-ai-social"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dinesh Pinto", email = "annual.fallout_0z@icloud.com" }
]
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.1",
    "google-generativeai>=0.8.4",
    "httpx>=0.28.1",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "pyopenssl>=25.0.0",
    "requests>=2.32.3",
    "structlog>=25.1.0",
]

[project.scripts]
start-social = "flare_ai_social.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.394",
    "ruff>=0.9.6",
]
tuning = [
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D203", "D212", "COM812", "D", "S105", "ANN401", "T201", "ISC003"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]
"src/flare_ai_social/prompts/dataset.py" = ["E501"]
"src/flare_ai_social/prompts/templates.py" = ["E501"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
pythonVersion = "3.12"
strictListInference = true
strictDictionaryInference = true
strictSetInference = true
deprecateTypingAliases = true
disableBytesTypePromotions = true
reportUnusedClass = true
reportUnusedFunction = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportPrivateUsage = true
reportPrivateImportUsage = false
reportTypeCommentUsage = true
reportConstantRedefinition = true
reportDeprecated = true
reportInconsistentConstructor = true
reportUninitializedInstanceVariable = true
reportMissingParameterType = true
reportCallInDefaultInitializer = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportInvalidStubStatement = true
reportUnusedExpression = true
reportUnnecessaryTypeIgnoreComment = true
reportMatchNotExhaustive = true
reportImplicitOverride = true
reportShadowedImports = true
